refactor(kernel): extract webhook-emitter and post-check from step-worker (GH-473)
Split step-worker.js (1747 lines) into three focused modules:
- webhook-emitter.js (138 lines): Thyra-aligned webhook envelope + retry
- post-check.js (254 lines): scope guard, contract validation, auto-finalize
- step-worker.js (1400 lines): core execution + lock management + prompt building

All existing require('./step-worker') paths remain unchanged — step-worker.js
re-exports everything from the sub-modules for backward compatibility.
